# docs/api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: AssuranceNet Document Management API
  description: |
    Azure-native document management API replacing Oracle Universal Content Manager (UCM)
    for FSIS AssuranceNet. Provides document upload, download, PDF conversion tracking,
    PDF merging, investigation management, and audit logging.

    ## FSIS Data Context
    This API manages food safety investigation documents from the
    [FSIS Science & Data program](https://www.fsis.usda.gov/science-data), including
    sampling plans, laboratory data (CSV), National Residue Program reports,
    microbiology baselines, MPI establishment directories, and inspection task records.

    ## Document Versioning
    Documents support versioning for compliance and audit requirements. Each document
    has a stable logical identity (document ID) and one or more versions (physical binaries).
    End users see only the latest version; administrators can access historical versions
    and rollback if needed.

    ## PDF Conversion
    The system uses Aspose.Words as the default PDF engine for document conversions.
    PDF conversion is tracked asynchronously; check `pdf_conversion_status` for
    progress (pending, processing, completed, failed, not_required).

    ## Merge Ordering
    PDF merges follow type-based ordering (FSIS-specific rules):
    1. Investigation Reports (10)
    2. Inspection Forms (20)
    3. Laboratory Results (30)
    4. Legal Documents (40)
    5. Correspondence (50)
...

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Microsoft Entra ID JWT access token
  schemas:
...
    DocumentVersion:
      type: object
      description: Immutable physical binary — one version of a logical document.
      properties:
        id:
          type: string
          format: uuid
          description: Version ID
        document_id:
          type: string
          format: uuid
          description: Parent document ID
        version_number:
          type: integer
          description: Sequential version number (1, 2, 3, ...)
        original_filename:
          type: string
          description: Uploaded filename
        mime_type:
          type: string
          nullable: true
          example: "application/pdf"
        file_size_bytes:
          type: integer
          format: int64
          description: Original file size
        checksum:
          type: string
          description: SHA-256 checksum of original file
        is_latest:
          type: boolean
          description: Whether this is the current/latest version
        pdf_conversion_status:
          type: string
          enum: [pending, processing, completed, failed, not_required]
          description: PDF conversion state
        pdf_conversion_error:
          type: string
          nullable: true
          description: Error details if conversion failed
        pdf_converted_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when PDF conversion completed
        scan_status:
          type: string
          enum: [pending, scanning, clean, infected, error]
          description: Malware scan state
        scanned_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when scan completed
        uploaded_by:
          type: string
          description: User ID who uploaded this version
        uploaded_by_name:
          type: string
          nullable: true
          description: Display name of uploader
        uploaded_at:
          type: string
          format: date-time
          description: Timestamp when uploaded

    # ========================================================================
    # Document Schema (Logical Identity + Latest Version)
    # ========================================================================
    Document:
      type: object
      description: |
        Logical document with stable identity and versioning support.
        End users see the latest version only; inlined latest version fields
        for convenience.
      properties:
        id:
          type: string
          format: uuid
          description: Document ID (stable across versions)
        investigation_id:
          type: string
          format: uuid
          description: Parent investigation ID
        document_type:
          type: string
          enum:
            - investigation_report
            - inspection_form
            - laboratory_result
            - correspondence
            - supporting_evidence
            - legal_document
            - other
          description: FSIS document classification
        title:
          type: string
          nullable: true
          description: Optional document title (e.g., "Lab Report for Sample XYZ")
        created_by:
          type: string
          description: User ID who created the document
        created_by_name:
          type: string
          nullable: true
          description: Display name of creator
        created_at:
          type: string
          format: date-time
          description: Document creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last updated timestamp
        # Latest version fields (inlined)
        current_version_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the current (latest) version
        version_number:
          type: integer
          nullable: true
          description: Current version number
        original_filename:
          type: string
          nullable: true
          description: Filename of current version
        mime_type:
          type: string
          nullable: true
          example: "application/pdf"
        file_size_bytes:
          type: integer
          format: int64
          nullable: true
          description: File size of current version
        checksum:
          type: string
          nullable: true
          description: SHA-256 of current version
        pdf_conversion_status:
          type: string
          nullable: true
          enum: [pending, processing, completed, failed, not_required]
        uploaded_by:
          type: string
          nullable: true
          description: User ID who uploaded current version
        uploaded_by_name:
          type: string
          nullable: true
          description: Display name of uploader
        uploaded_at:
          type: string
          format: date-time
          nullable: true
          description: Upload timestamp of current version

    # ========================================================================
    # Document Upload Response
    # ========================================================================
    DocumentUploadResponse:
      type: object
      description: Response after successfully uploading a document or new version.
      properties:
        document_id:
          type: string
          format: uuid
          description: The document ID
        version_id:
          type: string
          format: uuid
          description: The new version ID
        version_number:
          type: integer
          description: Version number (1 for new documents, increments for updates)
        original_filename:
          type: string
        file_size_bytes:
          type: integer
          format: int64
        checksum:
          type: string
          description: SHA-256 checksum
        document_type:
          type: string
          enum:
            - investigation_report
            - inspection_form
            - laboratory_result
            - correspondence
            - supporting_evidence
            - legal_document
            - other
        pdf_conversion_status:
          type: string
          enum: [pending, processing, completed, failed, not_required]
        blob_path:
          type: string
          description: Blob storage path of the uploaded file

    # ========================================================================
    # Rollback Response
    # ========================================================================
    RollbackResponse:
      type: object
      description: Response after rolling back a document to a previous version.
      properties:
        document_id:
          type: string
          format: uuid
        rolled_back_version:
          type: integer
          description: The version number that was rolled back (no longer current)
        promoted_version:
          type: integer
          description: The version number that was promoted to current
        new_current_version_id:
          type: string
          format: uuid
          description: The version ID now marked as current

    # ========================================================================
    # Admin Document Detail (with all versions)
    # ========================================================================
    AdminDocumentDetail:
      type: object
      description: Full document detail with all versions — admin only.
      properties:
        id:
          type: string
          format: uuid
        investigation_id:
          type: string
          format: uuid
        document_type:
          type: string
          enum:
            - investigation_report
            - inspection_form
            - laboratory_result
            - correspondence
            - supporting_evidence
            - legal_document
            - other
        title:
          type: string
          nullable: true
        created_by:
          type: string
        created_by_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        current_version_id:
          type: string
          format: uuid
          nullable: true
        is_deleted:
          type: boolean
        versions:
          type: array
          items:
            $ref: "#/components/schemas/DocumentVersion"
          description: All versions in descending order

    # ========================================================================
    # Health & Pagination
    # ========================================================================
...
paths:
...
  /api/v1/documents/{document_id}:
    get:
      summary: Get document metadata (latest version)
      tags: [Documents]
      description: Returns the current (latest) version metadata inlined.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
    delete:
      summary: Soft-delete document (all versions)
      tags: [Documents]
      description: Marks the document and all its versions as deleted.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Document deleted successfully
        "404":
          description: Document not found
...
  /api/v1/documents/{document_id}/versions:
    post:
      summary: Upload new version of existing document
      tags: [Documents]
      description: |
        Creates a new version (incremented version number) of an existing document.
        The new version becomes the current/latest version automatically.

        Form fields:
        - file: The new file (required)
        - title: Optional updated title
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  description: Optional updated title
      responses:
        "201":
          description: New version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
        "404":
          description: Document not found
        "413":
          description: File too large

  # ========================================================================
  # PDF Merge Endpoint
  # ========================================================================
...
  /api/v1/admin/documents/{document_id}:
    get:
      summary: Get document with all versions (Admin only)
      tags: [Admin - Documents]
      description: Returns complete document detail including all versions, deleted status, etc.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full document detail with version history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDocumentDetail"
        "403":
          description: Admin role required
        "404":
          description: Document not found
  /api/v1/admin/documents/{document_id}/versions:
    get:
      summary: List all versions of a document (Admin only)
      tags: [Admin - Documents]
      description: Returns all versions including historical ones, sorted descending by version number.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of all versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
        "403":
          description: Admin role required
        "404":
          description: Document not found
  /api/v1/admin/documents/{document_id}/versions/{version_id}/download:
    get:
      summary: Download specific version (Admin only)
      tags: [Admin - Documents]
      description: Downloads the original file for a specific historical version.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Admin role required
        "404":
          description: Document or version not found
  /api/v1/admin/documents/{document_id}/rollback:
    post:
      summary: Rollback document to a previous version (Admin only)
      tags: [Admin - Documents]
      description: |
        Marks a previous version as current (is_latest=True) and demotes the current
        version (is_latest=False). Enables recovery from accidental overwrite or corruption.
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_id]
              properties:
                version_id:
                  type: string
                  format: uuid
                  description: The version ID to promote as current
      responses:
        "200":
          description: Rollback successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollbackResponse"
        "400":
          description: Invalid request (version_id is already current, etc.)
        "403":
          description: Admin role required
        "404":
          description: Document or version not found

  # ========================================================================
  # Audit Endpoints
  # ========================================================================
...